Prevent simultaneous SchedulerShells from running
There may be some situations where a SchedulerShell is running a task which takes longer to execute than the amount of time between a new SchedulerShell being launched by your cron job. This will prevent the SchedulerShell starting again on the same task.

It can be disabled or adjusted with a configure setting.